fix(controller): 统一用户名长度，在注册和修改用户名时都限制长度为3-20位 - 统一 status 字段的类型

+++ LinkCloud/controller/stats.go
@@ -37,7 +37,7 @@ type accessLogRow struct {
 	DeviceType string    `gorm:"column:device_type"`
 	Referer    string    `gorm:"column:referer"`
 	UserAgent  string    `gorm:"column:user_agent"`
-	Status     int       `gorm:"column:status"`
+	Status     int8      `gorm:"column:status"`
 	ErrorMsg   string    `gorm:"column:error_message"`
 }
